👔 Merge /store & /store/books into one page
