+++ src/pages/store/featured.vue
